iot-demo-docker/backend/requirements.txt
flask==3.0.2
paho-mqtt==2.1.0
requests==2.31.0
dnspython==2.2.1 
python-socketio==5.9.0
python-engineio==4.7.0
gunicorn==21.2.0
flask-cors==4.0.0
Flask-SocketIO==5.3.4
eventlet==0.33.3
